refactor: Negotiation entity
